Templating of game rules and game options
