+++ mods/midnight/lib/midnight-game-rules.template.js
@@ -1,0 +1,16 @@
+module.exports = (app, mod) => {
+
+  let html = `<div class="rules-overlay">
+    <h1>Midnight Rogue</h1>
+    <blockquote>
+    Danger lurks in every corner of Port Blacksand — infamous city of evil-doers!
+    YOU are an apprentice in the Thieves' Guild of Port Blacksand. Tonight is the testing time, the climax of your training. Your mission is to find and steal a priceless gem, the Eye of the Basilisk, and the special skills you have learned will be tested to the limit!
+    Many terrors lie in wait in the darkness of this evil city's backstreets and alleys. Will you be able to accomplish your dangerous task? You have until morning to prove yourself.
+    </blockquote>
+    <h2>Specific Instructions</h2>
+    <p>You may carry up to six items with you in your rucksack, with the provisions counting together as one item. You are also gifted with abilities taught by the Guild of Thieves.</p>
+  </div>`;
+  return html;
+
+}
+
