+++ mods/spider/spider.js
@@ -1,5 +1,7 @@
 var saito = require('../../lib/saito/saito');
 var GameTemplate = require('../../lib/templates/gametemplate');
+const SpiderGameRulesTemplate = require("./lib/spider-game-rules.template");
+const SpiderGameOptionsTemplate = require("./lib/spider-game-options.template");
 
 
 //////////////////
@@ -34,31 +36,12 @@ class Spider extends GameTemplate {
   
 
   returnGameRulesHTML(){
-    return `<div class="rules-overlay">
-            <h1>Spider Saitolaire</h1>
-            <ul>
-            <li>You have ten slots in which to arrange two decks of playing cards. </li>
-            <li>Only half the cards are dealt at the beginning, and additional draws place a new card on each stack</li>
-            <li>Cards must be placed in numerical order, e.g. only a 3 can be placed on top of a 4.</li>
-            <li>Sequences of arranged cards of a single suit may be moved as a unit to another pile.</li>
-            <li>Any card may be placed on an open slot.</li>
-            <li>When you complete a set from A to K, it is immediately removed from gameplay</li>
-            </ul>
-            </div>
-            `;
-
+    return SpiderGameRulesTemplate(app, this);
   }
 
 
   returnSingularGameOption(app){
-    let saved_dif = app?.options?.gameprefs?.spider_difficulty || "medium";
-    console.log(saved_dif);
-    let html = `<select name="difficulty">
-              <option value="easy">Easy (1 suit) </option>
-              <option value="medium">Medium (2 suits) </option>
-              <option value="hard">Hard (4 suits) </option>
-            </select>`;
-    return html.replace(`${saved_dif}"`,`${saved_dif}" selected`);
+    return SpiderGameOptionsTemplate(app, this);
   }
     
   //Single player games don't allow game-creation and options prior to join
