+++ mods/thirteen/thirteen.js
@@ -1,5 +1,6 @@
 const GameTemplate = require('../../lib/templates/gametemplate');
-
+const ThirteenGameRulesTemplate = require("./lib/thirteen-game-rules.template");
+const ThirteenGameOptionsTemplate = require("./lib/thirteen-game-options.template");
 
 //////////////////
 // constructor  //
@@ -4399,38 +4400,12 @@ console.log("SHOULD PLACE: " + player);
 
 
   returnGameRulesHTML(){
-    return `<div class="rules-overlay">
-    <h1>Thirteen Days</h1>
-    <p>Thirteen Days is a fast paced two player simulation of the Cuban Missile Crisis. (No historical knowledge needed to play)</p>
-    <p>The game is played over three rounds, where each player has a secret AGENDA that awards them additional PRESTIGE at the end of the round for completing. Each player starts the round with 5 STRATEGY cards that may either be played for EVENTS or for COMMAND.</p>
-    <p>EVENTS are associated with the US, the USSR, or the UN (neutral). Playing a STRATEGY card associated with your opponent allows them to trigger the event, while you can only play it for COMMAND, when you may add/remove INFLUENCE cubes (up to the amount specified on the card) from a single battleground. You may only have up to 17 INFLUENCE on the board and cannot directly move it (i.e. you have to remove influence and add influence in two separate moves). Adding or removing more than one influence affects the DEFCON level, and battlegrounds may contain at most 5 INFLUENCE cubes.</p>
-    <p>If you hold the PERSONAL LETTER, you may play it along with another card for +1 INFLUENCE on the COMMAND.</p>
-    <p>The round ends after each player has played 4 STRATEGY cards. The remaining STRATEGY card is saved for the AFTERMATH. Players with the most influence in any world opinion battleground may utilize the special abilities (e.g. being able to escalate/de-escalate on DEFCON track, receiving the PERSONAL LETTER, or peeking at the top STRATEGY card and either discarding it or adding it to the AFTERMATH). The secret AGENDAs are then revealed to see how players change their PRESTIGE. A player may only lead by 5 PRESTIGE, but this does not trigger an automatic win.</p>
-    <p>At the end of the round, if you have any markers in the DEFCON 1 zone or more than one in the DEFCON 2 zone, you trigger a thermonuclear war and lose the game. DEFCON is advanced by one for all markers at the beginning of every round.</p>
-    <p>After three rounds, the AFTERMATH cards are revealed and the player with the most INFLUENCE cubes associated with the cards receives an additional +2 PRESTIGE. The player with the most PRESTIGE wins the game.</p>
-    </div>`;
+    return ThirteenGameRulesTemplate(app, this);
   }
-
-
 
   returnGameOptionsHTML() {
-
-    return `
-            <h1 class="overlay-title">Thirteen Days</h1>
-            <div class="overlay-input">
-            <label for="player1">Play as:</label>
-            <select name="player1">
-              <option value="random">random</option>
-              <option value="ussr" default>USSR</option>
-              <option value="us">US</option>
-            </select>
-            </div>
-          `;
-
+    return ThirteenGameOptionsTemplate(app, this);
   }
-
-
-
 
   returnUnplayedCards() {
