+++ mods/twilight/lib/twilight-game-options-template.js
@@ -1,0 +1,226 @@
+module.exports = (app, mod) => {
+
+  let html =  `
+
+      <div style="padding:40px;width:100vw;height:100vh;overflow-y:scroll;display:grid;grid-template-columns: 200px auto">
+
+  <div style="top:0;left:0;">
+
+            
+
+            <label for="deck">Deck:</label>
+            <select name="deck" id="deckselect" onchange='
+        if ($("#deckselect").val() == "saito") { 
+    $(".saito_edition").prop("checked",true); 
+    $(".endofhistory_edition").prop("checked", false); 
+        } else { 
+    $(".saito_edition").prop("checked", false); 
+          if ($("#deckselect").val() == "optional") { 
+      $(".optional_edition").prop("checked", false); 
+    } else { 
+      $(".optional_edition").prop("checked", true); 
+      if ($("#deckselect").val() == "endofhistory") { 
+        $(".endofhistory_edition").prop("checked",true); 
+        $(".optional_edition").prop("checked", false);
+      } else {
+        if ($("#deckselect").val() == "coldwarcrazies") { 
+          $(".coldwarcrazies_edition").prop("checked",true); 
+          $(".optional_edition").prop("checked", false);
+        } else {
+          if ($("#deckselect").val() == "absurdum") { 
+            $(".absurdum_edition").prop("checked",true); 
+            $(".optional_edition").prop("checked",true);
+          }
+        }
+      }
+    }
+        } '>
+            <option value="original">original</option>
+              <option value="optional" selected>optional</option>
+              <option value="late-war">late war</option>
+              <option value="saito">saito edition</option>
+              <option value="absurdum">twilight absurdum</option>
+              <option value="endofhistory">end of history</option>
+              <option value="coldwarcrazies">cold war crazies</option>
+            </select>
+
+            <label for="usbonus">US bonus: </label>
+            <select name="usbonus">
+              <option value="0">0</option>
+              <option value="1">1</option>
+              <option value="2" selected>2</option>
+              <option value="3">3</option>
+              <option value="4">4</option>
+              <option value="5">5</option>
+              <option value="6">6</option>
+              <option value="7">7</option>
+              <option value="8">8</option>
+              <option value="9">9</option>
+              <option value="10">10</option>
+            </select>
+
+            <label for="clock">Player Time Limit:</label>
+            <select name="clock">
+              <option value="0" default>no limit</option>
+              <option value="10">10 minutes</option>
+              <option value="20">20 minutes</option>
+              <option value="30">30 minutes</option>
+              <option value="60">60 minutes</option>
+              <option value="90">90 minutes</option>
+              <option value="120">120 minutes</option>
+            </select>
+
+            <label for="observer_mode">Observer Mode:</label>
+            <select name="observer">
+              <option value="enable" >enable</option>
+              <option value="disable" selected>disable</option>
+            </select>
+            <div id="game-wizard-advanced-return-btn" class="game-wizard-advanced-return-btn button">accept</div>
+  </div>
+
+            <div id="game-wizard-advanced-box" class="game-wizard-advanced-box" style="display:block;padding-left:20px;">
+
+        <style type="text/css">li { list-style: none; } .saito-select { margin-bottom: 10px; margin-top:5px; } label { text-transform: uppercase; } .removecards { grid-gap: 0.1em; } .list-header { font-weight: bold; font-size:1.5em; margin-top:0px; margin-bottom:10px; margin-left: 15px; text-transform: uppercase; } </style>
+              <div class="list-header">remove cards:</div>
+              <ul id="removecards" class="removecards">
+              <li><input class="remove_card" type="checkbox" name="asia" /> Asia Scoring</li>
+              <li><input class="remove_card" type="checkbox" name="europe" /> Europe Scoring</li>
+              <li><input class="remove_card" type="checkbox" name="mideast" /> Middle-East Scoring</li>
+              <li><input class="remove_card" type="checkbox" name="duckandcover" /> Duck and Cover</li>
+              <li><input class="remove_card" type="checkbox" name="fiveyearplan" /> Five Year Plan</li>
+              <li><input class="remove_card" type="checkbox" name="socgov" /> Socialist Governments</li>
+              <li><input class="remove_card" type="checkbox" name="fidel" /> Fidel</li>
+              <li><input class="remove_card" type="checkbox" name="vietnamrevolts" /> Vietnam Revolts</li>
+              <li><input class="remove_card" type="checkbox" name="blockade" /> Blockade</li>
+              <li><input class="remove_card" type="checkbox" name="koreanwar" /> Korean War</li>
+              <li><input class="remove_card" type="checkbox" name="romanianab" /> Romanian Abdication</li>
+              <li><input class="remove_card" type="checkbox" name="arabisraeli" /> Arab Israeli War</li>
+              <li><input class="remove_card" type="checkbox" name="comecon" /> Comecon</li>
+              <li><input class="remove_card" type="checkbox" name="nasser" /> Nasser</li>
+              <li><input class="remove_card" type="checkbox" name="warsawpact" /> Warsaw Pact</li>
+              <li><input class="remove_card" type="checkbox" name="degaulle" /> De Gaulle Leads France</li>
+              <li><input class="remove_card" type="checkbox" name="naziscientist" /> Nazi Scientists Captured</li>
+              <li><input class="remove_card" type="checkbox" name="truman" /> Truman</li>
+              <li><input class="remove_card saito_edition" type="checkbox" name="olympic" /> Olympic Games</li>
+              <li><input class="remove_card" type="checkbox" name="nato" /> NATO</li>
+              <li><input class="remove_card" type="checkbox" name="indreds" /> Independent Reds</li>
+              <li><input class="remove_card" type="checkbox" name="marshall" /> Marshall Plan</li>
+              <li><input class="remove_card" type="checkbox" name="indopaki" /> Indo-Pakistani War</li>
+              <li><input class="remove_card" type="checkbox" name="containment" /> Containment</li>
+              <li><input class="remove_card" type="checkbox" name="cia" /> CIA Created</li>
+              <li><input class="remove_card" type="checkbox" name="usjapan" /> US/Japan Defense Pact</li>
+              <li><input class="remove_card" type="checkbox" name="suezcrisis" /> Suez Crisis</li>
+              <li><input class="remove_card" type="checkbox" name="easteuropean" /> East European Unrest</li>
+              <li><input class="remove_card" type="checkbox" name="decolonization" /> Decolonization</li>
+              <li><input class="remove_card" type="checkbox" name="redscare" /> Red Scare</li>
+              <li><input class="remove_card" type="checkbox" name="unintervention" /> UN Intervention</li>
+              <li><input class="remove_card" type="checkbox" name="destalinization" /> Destalinization</li>
+              <li><input class="remove_card" type="checkbox" name="nucleartestban" /> Nuclear Test Ban Treaty</li>
+              <li><input class="remove_card" type="checkbox" name="formosan" /> Formosan Resolution</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="defectors" /> Defectors</li>
+              <li><input class="remove_card optional_edition " type="checkbox" name="specialrelation" /> Special Relationship</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="cambridge" /> The Cambridge Five</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="norad" /> NORAD</li>
+            </ul>
+            <ul class="removecards" style="clear:both;margin-top:13px">
+              <li><input class="remove_card" type="checkbox" name="brushwar" /> Brush War</li>
+              <li><input class="remove_card" type="checkbox" name="camerica" /> Central America Scoring</li>
+              <li><input class="remove_card" type="checkbox" name="seasia" /> Southeast Asia Scoring</li>
+              <li><input class="remove_card" type="checkbox" name="armsrace" /> Arms Race</li>
+              <li><input class="remove_card" type="checkbox" name="cubanmissile" /> Cuban Missile Crisis</li>
+              <li><input class="remove_card" type="checkbox" name="nuclearsubs" /> Nuclear Subs</li>
+              <li><input class="remove_card" type="checkbox" name="quagmire" /> Quagmire</li>
+              <li><input class="remove_card" type="checkbox" name="saltnegotiations" /> Salt Negotiations</li>
+              <li><input class="remove_card" type="checkbox" name="beartrap" /> Bear Trap</li>
+              <li><input class="remove_card saito_edition" type="checkbox" name="summit" /> Summit</li>
+              <li><input class="remove_card" type="checkbox" name="howilearned" /> How I Learned to Stop Worrying</li>
+              <li><input class="remove_card" type="checkbox" name="junta" /> Junta</li>
+              <li><input class="remove_card" type="checkbox" name="kitchendebates" /> Kitchen Debates</li>
+              <li><input class="remove_card" type="checkbox" name="missileenvy" /> Missile Envy</li>
+              <li><input class="remove_card" type="checkbox" name="wwby" /> We Will Bury You</li>
+              <li><input class="remove_card" type="checkbox" name="brezhnev" /> Brezhnev Doctrine</li>
+              <li><input class="remove_card" type="checkbox" name="portuguese" /> Portuguese Empire Crumbles</li>
+              <li><input class="remove_card" type="checkbox" name="southafrican" /> South African Unrest</li>
+              <li><input class="remove_card" type="checkbox" name="allende" /> Allende</li>
+              <li><input class="remove_card" type="checkbox" name="willybrandt" /> Willy Brandt</li>
+              <li><input class="remove_card" type="checkbox" name="muslimrevolution" /> Muslim Revolution</li>
+              <li><input class="remove_card" type="checkbox" name="abmtreaty" /> ABM Treaty</li>
+              <li><input class="remove_card" type="checkbox" name="culturalrev" /> Cultural Revolution</li>
+              <li><input class="remove_card" type="checkbox" name="flowerpower" /> Flower Power</li>
+              <li><input class="remove_card" type="checkbox" name="u2" /> U-2 Incident</li>
+              <li><input class="remove_card" type="checkbox" name="opec" /> OPEC</li>
+              <li><input class="remove_card" type="checkbox" name="lonegunman" /> Lone Gunman</li>
+              <li><input class="remove_card" type="checkbox" name="colonial" /> Colonial</li>
+              <li><input class="remove_card" type="checkbox" name="panamacanal" /> Panama Canal</li>
+              <li><input class="remove_card" type="checkbox" name="campdavid" /> Camp David Accords</li>
+              <li><input class="remove_card" type="checkbox" name="puppet" /> Puppet Governments</li>
+              <li><input class="remove_card" type="checkbox" name="grainsales" /> Grain Sales to Soviets</li>
+              <li><input class="remove_card" type="checkbox" name="johnpaul" /> John Paul</li>
+              <li><input class="remove_card" type="checkbox" name="deathsquads" /> Death Squads</li>
+              <li><input class="remove_card" type="checkbox" name="oas" /> OAS Founded</li>
+              <li><input class="remove_card" type="checkbox" name="nixon" /> Nixon Plays the China Card</li>
+              <li><input class="remove_card" type="checkbox" name="sadat" /> Sadat Expels Soviets</li>
+              <li><input class="remove_card" type="checkbox" name="shuttle" /> Shuttle Diplomacy</li>
+              <li><input class="remove_card" type="checkbox" name="voiceofamerica" /> Voice of America</li>
+              <li><input class="remove_card" type="checkbox" name="liberation" /> Liberation Theology</li>
+              <li><input class="remove_card" type="checkbox" name="ussuri" /> Ussuri River Skirmish</li>
+              <li><input class="remove_card" type="checkbox" name="asknot" /> Ask Not What Your Country Can Do For You</li>
+              <li><input class="remove_card" type="checkbox" name="alliance" /> Alliance for Progress</li>
+              <li><input class="remove_card" type="checkbox" name="africa" /> Africa Scoring</li>
+              <li><input class="remove_card" type="checkbox" name="onesmallstep" /> One Small Step</li>
+              <li><input class="remove_card" type="checkbox" name="samerica" /> South America</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="che" /> Che</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="tehran" /> Our Man in Tehran</li>
+            </ul>
+            <ul class="removecards" style="clear:both;margin-top:13px">
+              <li><input class="remove_card" type="checkbox" name="iranianhostage" /> Iranian Hostage Crisis</li>
+              <li><input class="remove_card" type="checkbox" name="ironlady" /> The Iron Lady</li>
+              <li><input class="remove_card" type="checkbox" name="reagan" /> Reagan Bombs Libya</li>
+              <li><input class="remove_card" type="checkbox" name="starwars" /> Star Wars</li>
+              <li><input class="remove_card" type="checkbox" name="northseaoil" /> North Sea Oil</li>
+              <li><input class="remove_card" type="checkbox" name="reformer" /> The Reformer</li>
+              <li><input class="remove_card" type="checkbox" name="marine" /> Marine Barracks Bombing</li>
+              <li><input class="remove_card" type="checkbox" name="KAL007" /> Soviets Shoot Down KAL-007</li>
+              <li><input class="remove_card" type="checkbox" name="glasnost" /> Glasnost</li>
+              <li><input class="remove_card" type="checkbox" name="ortega" /> Ortega Elected in Nicaragua</li>
+              <li><input class="remove_card" type="checkbox" name="terrorism" /> Terrorism</li>
+              <li><input class="remove_card" type="checkbox" name="ironcontra" /> Iran Contra Scandal</li>
+              <li><input class="remove_card" type="checkbox" name="chernobyl" /> Chernobyl</li>
+              <li><input class="remove_card" type="checkbox" name="debtcrisis" /> Latin American Debt Crisis</li>
+              <li><input class="remove_card" type="checkbox" name="teardown" /> Tear Down this Wall</li>
+              <li><input class="remove_card" type="checkbox" name="evilempire" /> An Evil Empire</li>
+              <li><input class="remove_card" type="checkbox" name="aldrichames" /> Aldrich Ames Remix</li>
+              <li><input class="remove_card" type="checkbox" name="pershing" /> Pershing II Deployed</li>
+              <li><input class="remove_card" type="checkbox" name="wargames" /> Wargames</li>
+              <li><input class="remove_card" type="checkbox" name="solidarity" /> Solidarity</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="iraniraq" /> Iran-Iraq War</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="yuri" /> Yuri and Samantha</li>
+              <li><input class="remove_card optional_edition" type="checkbox" name="awacs" /> AWACS Sale to Saudis</li>
+            </ul>
+
+            <div class="list-header">add cards:</div>
+            <ul id="removecards" class="removecards">
+              <li><input class="remove_card saito_edition" type="checkbox" name="culturaldiplomacy" /> Cultural Diplomacy (Early-War)</li>
+              <li><input class="remove_card saito_edition" type="checkbox" name="handshake" /> Handshake in Space (Mid-War)</li>
+              <li><input class="remove_card saito_edition" type="checkbox" name="rustinredsquare" /> Rust Lands in Red Square (Late-War)</li>
+              <li><input class="remove_card" type="checkbox" name="gouzenkoaffair" /> Gouzenko Affair (Early-War)</li>
+              <li><input class="remove_card" type="checkbox" name="poliovaccine" /> Polio Vaccine (Early-War)</li>
+              <li><input class="remove_card saito_edition" type="checkbox" name="berlinagreement" /> 1971 Berlin Agreement (Mid-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="peronism" /> Peronism (Early-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="manwhosavedtheworld" /> The Man Who Saved the World (Mid-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="breakthroughatlopnor" /> Breakthrough at Lop Nor (Mid-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="nationbuilding" /> Nation Building (Mid-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="greatsociety" /> Great Society (Mid-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="perestroika"  /> Perestroika (Late-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="eurocommunism" /> Eurocommunism (Mid-War)</li>
+              <li><input class="remove_card endofhistory_edition" type="checkbox" name="inftreaty" /> INF Treaty (Late-War)</li>
+              <li><input class="remove_card coldwarcrazies_edition" type="checkbox" name="communistrevolution" /> Communist Revolution (Early-War)</li>
+            </div>
+
+      </div>
+    </div>
+          `;
+
+    return html;
+
+}
