+++ mods/twilight/lib/twilight-singular-game-options-template.js
@@ -1,0 +1,12 @@
+module.exports = (app, mod) => {
+
+  let html =  `<div><label for="player1">Play as:</label>
+            <select name="player1">
+              <option value="random" selected>random</option>
+              <option value="ussr">USSR</option>
+              <option value="us">US</option>
+            </select></div>
+          `;
+    return html;
+
+}
